creates logger file from the logger menu. logger menu is now accessible from full  namer.
