+++ build.gradle
@@ -17,8 +17,8 @@ repositories {
 
 dependencies {
 	testCompile(
-		[group: 'org.apache.poi', name: 'poi', version: '4.0.0'],
-		[group: 'org.apache.poi', name: 'poi-ooxml', version: '4.0.0'],
+		[group: 'org.apache.poi', name: 'poi', version: '3.17'],
+		[group: 'org.apache.poi', name: 'poi-ooxml', version: '3.17'],
 		[group: 'junit', name: 'junit', version: '4.12'],
 		[group: 'com.jfoenix', name: 'jfoenix', version: '8.0.2'],
 		[group: 'org.apache.derby', name: 'derby', version: '10.11.1.1'],
@@ -27,8 +27,8 @@ dependencies {
 		[group: 'org.mindrot', name: 'jbcrypt', version: '0.3m'],
 	)
 	compile(
-        [group: 'org.apache.poi', name: 'poi', version: '4.0.0'],
-        [group: 'org.apache.poi', name: 'poi-ooxml', version: '4.0.0'],
+        [group: 'org.apache.poi', name: 'poi', version: '3.17'],
+        [group: 'org.apache.poi', name: 'poi-ooxml', version: '3.17'],
         [group: 'junit', name: 'junit', version: '4.12'],
 		[group: 'com.jfoenix', name: 'jfoenix', version: '8.0.2'],
 		[group: 'org.apache.derby', name: 'derby', version: '10.11.1.1'],
